+++ lib/jeckle/model.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
 module Jeckle
-  module Model
-    def self.included(base)
-      base.include ActiveModel::Validations
-      base.include Virtus.model
-    end
+  class Model < Dry::Struct
+    transform_keys(&:to_sym)
+    transform_types(&:omittable)
+
+    include ActiveModel::Validations
   end
 end